Move TestingAddresses to c.s.c.testing package
