+++ src/test/groovy/com/streamr/client/utils/DecryptionQueuesSpec.groovy
@@ -2,7 +2,7 @@ package com.streamr.client.utils
 
 import com.streamr.client.protocol.message_layer.StreamMessage
 import com.streamr.client.protocol.message_layer.StreamrSpecification
-import com.streamr.client.protocol.message_layer.TestingAddresses
+import com.streamr.client.testing.TestingAddresses
 
 class DecryptionQueuesSpec extends StreamrSpecification {
